SpeechLLm/requirements.txt
faster-whisper
torch
numpy
sounddevice
soundfile
pyaudio
pyyaml
python-dotenv
requests